Refactor SeriesLink with SectionLink fallback
